server/requirements.txt
diplomacy==1.1.2
python-dotenv==0.19.2 
web3==5.24.0
